Improve Providers and typesafety
