+++ enterprise/frontend/src/metabase-enterprise/data_apps/components/DataAppProvider.tsx
@@ -1,13 +1,15 @@
-import { Global } from "@emotion/react";
-import type { ReactNode } from "react";
+import createCache from "@emotion/cache";
+import { CacheProvider, Global } from "@emotion/react";
+import { type ReactNode, useMemo } from "react";
 
 import { SCOPED_CSS_RESET } from "embedding-sdk-bundle/components/private/PublicComponentStylesWrapper";
 import { PortalContainer } from "embedding-sdk-bundle/components/private/SdkPortalContainer";
 import { SdkThemeProvider } from "embedding-sdk-bundle/components/private/SdkThemeProvider";
 import type { MetabaseTheme } from "metabase/embedding-sdk/theme";
 import { MetabaseReduxProvider } from "metabase/redux";
+import { getCspNonce } from "metabase/utils/csp";
 
-import { getHostBackedSdkStore } from "../host-sdk-init";
+import { useHostSdkStore } from "../lib/use-host-sdk-store";
 
 // Note: Mantine + SDK CSS is loaded into the iframe via the `data-app-vendors`
 // Rspack entry (`<link rel="stylesheet">` in the iframe srcdoc). Don't side-
@@ -30,18 +32,29 @@ interface DataAppProviderProps {
  * Takes only `{ theme, children }` — no `authConfig`. The data-app surface
  * inherits the host's session cookie; auth is already established.
  */
-export const DataAppProvider = ({ theme, children }: DataAppProviderProps) => {
-  const sdkStore = getHostBackedSdkStore();
+export const DataAppProvider = (props: DataAppProviderProps) => {
+  const { children, ...restProps } = props;
+  const { theme } = restProps;
+  const sdkStore = useHostSdkStore(restProps);
+
+  // Iframe-side Emotion cache: keyed + CSP-nonced so SDK/Mantine styles inject
+  // into this document's head (not the parent's).
+  const emotionCache = useMemo(
+    () => createCache({ key: "data-app", nonce: getCspNonce() ?? undefined }),
+    [],
+  );
 
   return (
-    <MetabaseReduxProvider store={sdkStore}>
-      <SdkThemeProvider theme={theme}>
-        <Global styles={SCOPED_CSS_RESET} />
+    <CacheProvider value={emotionCache}>
+      <MetabaseReduxProvider store={sdkStore}>
+        <SdkThemeProvider theme={theme}>
+          <Global styles={SCOPED_CSS_RESET} />
 
-        {children}
+          {children}
 
-        <PortalContainer />
-      </SdkThemeProvider>
-    </MetabaseReduxProvider>
+          <PortalContainer />
+        </SdkThemeProvider>
+      </MetabaseReduxProvider>
+    </CacheProvider>
   );
 };
